requirements.txt
# Python dependencies for Network-IDS-ML

# Core ML libraries
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
imbalanced-learn>=0.9.0
scipy>=1.7.0

# Model persistence
joblib>=1.1.0

# Configuration management
PyYAML>=6.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Explainability (optional but recommended)
shap>=0.41.0

# Web API (for deployment)
flask>=2.0.0
requests>=2.26.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0
